+++ window_x11.c
@@ -26,7 +26,7 @@ bool running = true;
 bool mouseHeld = false;
 int currentSceneId = 0;
 
-float getTime() {
+double getTime() {
 	struct timespec ts;
 	clock_gettime(CLOCK_MONOTONIC, &ts);
 	return ts.tv_sec + ts.tv_nsec / 1e9;
